Adding GPIO, I2C and PWM tests

+++ module/motor/include/servo_motor.hpp
@@ -14,19 +14,19 @@ namespace motor {
 
 class ServoMotor : protected hal::pwm::PWM_Component {
  public:
-  ServoMotor(hal::pwm::PWM_ChannelId pwm_channel, float default_angle = 0.0F,
+  ServoMotor(hal::pwm::PWM_ChannelId pwm_channel, double default_angle = 0.0F,
              bool is_enabled = true);
   void enable() { enableOutput(); }
   void disable() { disableOutput(); }
-  void setAngle(float angle);
-  float getAngle() const noexcept { return current_angle_; }
+  void setAngle(double angle);
+  double getAngle() const noexcept { return current_angle_; }
   bool isEnabled() const noexcept { return PWM_Component::isEnabled(); }
 
  private:
-  uint32_t parseAngle(float angle);
+  uint32_t parseAngle(double angle);
 
  private:
-  float current_angle_;
+  double current_angle_;
 };
 
 }  // namespace motor
